requirements.txt
cycler
kiwisolve
matplotli
numpy
pyparsin
PyQt5
PyQt5-sip
pyqtgraph
python-dateuti
six
